feat: lex-sort query args and arg transformer hook
This commit introduces a new feature that enables the specification
of specific query parameters to be used as the cache index key
when creating a URI via a arg transformer hook.

For instance, it allows caching requests like GET /foo?mode=bar. The
assumption is that users are aware of the query parameters that can
be sent and that these parameters are utilized by the handler to modify
the request. Consequently, even if the user includes additional
irrelevant parameters like GET /foo?mode=bar&junk=xyz, the cache will
still be utilized since the specified parameters are the key.

Also, we now lexicographically sort the query params so that cache
cannot be busted by reordering the params.