feat: support specifying query params for caching
This commit introduces a new feature that enables the specification
of specific query parameters to be used as the cache index key
when creating a URI.

Users can now define a map of query parameters that will be taken
into account when caching requests.

For instance, it allows caching requests like GET /foo?mode=bar. The
assumption is that users are aware of the query parameters that can
be sent and that these parameters are utilized by the handler to modify
the request. Consequently, even if the user includes additional
irrelevant parameters like GET /foo?mode=bar&junk=xyz, the cache will
still be utilized since the specified parameters are the key.